Add initial functionality
Copied the design from alexpapworth/mytodolist.

Click on a Timekeeper to chnage the title. Check on the timestamp when you've completed the task.

Press the plus button for a new Timekeeper. Press the minus button to delete the Timekeeper at the bottom of the list.

Next up, need to order Timekeeprs by "recent completion".